+++ NEWS.md
@@ -1,4 +1,5 @@
 # pipfun (development version)
+* deprecate `path` in `log_save()`. Now we use pins boards. 
 
 # pipfun 0.3.6
 * Add tests and add functionality to retrieve specif boards in `get_pins_boards()`